refactor: changed x and y to double type

+++ src/main/java/com/hybris/homeserver/pictochat/message/Vector2.java
@@ -1,23 +1,23 @@
 package com.hybris.homeserver.pictochat.message;
 
 public class Vector2 {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
 	public Vector2() {
 		
 	}
 	
-	public Vector2(int x, int y) {
+	public Vector2(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return this.x;
 	}
 	
-	public int getY() {
+	public double getY() {
 		return this.y;
 	}
